function/requirements.txt
boto3~=1.24.89
requests~=2.28.1
psycopg2-binary~=2.9.3